Rename mr integration tests
