+++ integration-test/mr/src/test/java/org/apache/uniffle/test/SecondarySortIT.java
@@ -17,12 +17,13 @@
 
 package org.apache.uniffle.test;
 
-import java.util.List;
 import java.util.Map;
 import java.util.Random;
 
-import com.google.common.collect.Lists;
 import com.google.common.collect.Maps;
+import org.apache.hadoop.conf.Configurable;
+import org.apache.hadoop.conf.Configuration;
+import org.apache.hadoop.examples.SecondarySort;
 import org.apache.hadoop.fs.FSDataOutputStream;
 import org.apache.hadoop.fs.Path;
 import org.apache.hadoop.io.IntWritable;
@@ -30,7 +31,6 @@ import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapred.FileInputFormat;
 import org.apache.hadoop.mapred.FileOutputFormat;
 import org.apache.hadoop.mapred.JobConf;
-import org.apache.hadoop.mapred.WordCount;
 import org.apache.hadoop.mapreduce.Job;
 import org.apache.hadoop.mapreduce.RssMRConfig;
 import org.apache.hadoop.util.Tool;
@@ -41,12 +41,9 @@ import org.apache.uniffle.coordinator.CoordinatorConf;
 import org.apache.uniffle.server.ShuffleServerConf;
 import org.apache.uniffle.storage.util.StorageType;
 
-public class WordCountTest extends MRIntegrationTestBase {
+public class SecondarySortIT extends MRIntegrationTestBase {
 
-  String inputPath = "word_count_input";
-  List<String> wordTable = Lists.newArrayList("apple",
-      "banana", "fruit", "cherry", "Chinese", "America", "Japan",
-      "tomato");
+  String inputPath = "secondary_sort_input";
 
   @BeforeAll
   public static void setupServers() throws Exception {
@@ -62,9 +59,21 @@ public class WordCountTest extends MRIntegrationTestBase {
   }
 
   @Test
-  public void wordCountTest() throws Exception {
+  public void secondarySortTest() throws Exception {
     generateInputFile();
     run();
+  }
+
+  private void generateInputFile() throws Exception {
+    FSDataOutputStream outputStream = fs.create(new Path(inputPath));
+    Random random = new Random();
+    for (int i = 0; i < 100; i++) {
+      int first = random.nextInt();
+      int second = random.nextInt();
+      String str = first + " " + second + "\n";
+      outputStream.writeBytes(str);
+    }
+    outputStream.close();
   }
 
   @Override
@@ -72,40 +81,37 @@ public class WordCountTest extends MRIntegrationTestBase {
     return new TestTool();
   }
 
-  private void generateInputFile() throws Exception {
-    FSDataOutputStream outputStream = fs.create(new Path(inputPath));
-    Random random = new Random();
-    for (int i = 0; i < 100; i++) {
-      int index = random.nextInt(wordTable.size());
-      String str = wordTable.get(index) + "\n";
-      outputStream.writeBytes(str);
-    }
-    outputStream.close();
-  }
+  private class TestTool  extends SecondarySort implements Tool, Configurable {
 
-  private class TestTool extends WordCount {
-
-    String outputPath = "word_count_output/" + System.currentTimeMillis();
+    String outputPath = "secondary_sort_output/" + System.currentTimeMillis();
+    Configuration toolConf;
 
     @Override
-    public int run(String[] args) throws Exception {
+    public int run(String[] strings) throws Exception {
       JobConf conf = (JobConf) getConf();
-      conf.setJobName("wordcount");
-
-      // the keys are words (strings)
-      conf.setOutputKeyClass(Text.class);
-      // the values are counts (ints)
-      conf.setOutputValueClass(IntWritable.class);
-
-      conf.setMapperClass(MapClass.class);
-      conf.setCombinerClass(Reduce.class);
-      conf.setReducerClass(Reduce.class);
-      conf.setNumMapTasks(1);
-      conf.setNumReduceTasks(1);
       FileInputFormat.setInputPaths(conf, new Path(inputPath));
       FileOutputFormat.setOutputPath(conf, new Path(outputPath));
       Job job = new Job(conf);
+      job.setJarByClass(SecondarySort.class);
+      job.setMapperClass(SecondarySort.MapClass.class);
+      job.setReducerClass(SecondarySort.Reduce.class);
+      job.setPartitionerClass(SecondarySort.FirstPartitioner.class);
+      job.setGroupingComparatorClass(SecondarySort.FirstGroupingComparator.class);
+      job.setMapOutputKeyClass(SecondarySort.IntPair.class);
+      job.setMapOutputValueClass(IntWritable.class);
+      job.setOutputKeyClass(Text.class);
+      job.setOutputValueClass(IntWritable.class);
       return job.waitForCompletion(true) ? 0 : 1;
+    }
+
+    @Override
+    public void setConf(Configuration configuration) {
+      this.toolConf = configuration;
+    }
+
+    @Override
+    public Configuration getConf() {
+      return toolConf;
     }
   }
 }
